refactor(git): rename `repo` module
