+++ src/commands/init.rs
@@ -10,7 +10,7 @@ use fn_error_context::context;
 use log::warn;
 
 use crate::core::config::get_core_hooks_path;
-use crate::core::repo::Repo;
+use crate::git::Repo;
 use crate::util::{run_git_silent, wrap_git_error, GitRunInfo, GitVersion};
 
 const ALL_HOOKS: &[(&str, &str)] = &[
