+++ src/commands/move.rs
@@ -10,8 +10,8 @@ use crate::core::eventlog::{EventLogDb, EventReplayer};
 use crate::core::formatting::Glyphs;
 use crate::core::graph::{make_graph, BranchOids, CommitGraph, HeadOid, MainBranchOid};
 use crate::core::mergebase::MergeBaseDb;
-use crate::core::repo::Repo;
 use crate::core::rewrite::{execute_rebase_plan, ExecuteRebasePlanOptions, RebasePlanBuilder};
+use crate::git::Repo;
 use crate::util::{resolve_commits, GitRunInfo, ResolveCommitsResult};
 
 fn resolve_base_commit(graph: &CommitGraph, oid: git2::Oid) -> git2::Oid {
