+++ src/core.rs
@@ -6,6 +6,5 @@ pub mod formatting;
 pub mod graph;
 pub mod mergebase;
 pub mod metadata;
-pub mod repo;
 pub mod rewrite;
 pub mod tui;
