+++ src/git.rs
@@ -1,0 +1,5 @@
+//! Tools for interfacing with the Git repository.
+
+mod repo;
+
+pub use repo::Repo;
