+++ src/lib.rs
@@ -41,5 +41,6 @@
 
 pub mod commands;
 pub mod core;
+pub mod git;
 pub mod testing;
 pub mod util;
